requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.1.0
marshmallow==3.21.1
python-dotenv==1.0.1
pyodbc==5.0.1
SQLAlchemy==2.0.29


